Update caps.S
In order to have every lowercase character capitalized, I first loaded the address of the first character of the array into a register(r0). Then I set up r1 to be an offset, with which I traversed the array by setting it to 4 and adding it to the address of the current array character, because each character takes up four bits, having the offset equal to four ensures that the next element in the array will be found without trouble. Then the program loops through the array making a comparison to find if it is capital or lowercase then branching to the appropriate section based on the result.